requirements.txt
numpy
pandas
google-cloud-bigquery 
google-cloud-bigquery-storage 
pyarrow 
db-dtypes
pycaret
google-auth
seaborn
shap
scikit-learn==1.4.2